requirements-train.txt
# =============================================================================
# RAG 코어 + SFT 학습 — Python 3.10 / .venv
# vLLM 서빙은 반드시 별도 venv + requirements-vllm.txt
# =============================================================================
#
# [1] PyTorch 선설치 (CUDA에 맞게 하나만 선택):
#   CUDA 12.8:
#     pip install torch==2.10.0 torchvision==0.25.0 --index-url https://download.pytorch.org/whl/cu128
#   CUDA 12.1:
#     pip install torch==2.10.0 torchvision==0.25.0 --index-url https://download.pytorch.org/whl/cu121
#
# [2] 패키지 설치:
#     pip install -e .
#     pip install -r requirements-train.txt
#
# [3] MeCab (Ubuntu):
#     sudo apt-get install -y mecab libmecab-dev mecab-ipadic-utf8
#
# [4] Flash Attention 2 (선택, GPU 빌드 환경):
#     pip install flash-attn==2.6.3 --no-build-isolation
#
# [5] Qwen3.5 GatedDeltaNet 필수 (없으면 PyTorch fallback → OOM):
#     pip install causal-conv1d
#     pip install flash-linear-attention
# =============================================================================

# HuggingFace
# transformers 5.2.0+: Qwen3.5 지원 / unsloth 호환 상한 5.5.0
transformers==5.5.0
tokenizers==0.22.2
huggingface-hub>=1.0
accelerate==1.13.0
sentence-transformers==5.3.0
einops==0.8.0

# RAG / 벡터 / 희소
llama-index-core>=0.13.0,<0.15
llama-index-embeddings-huggingface==0.7.0
llama-index-vector-stores-qdrant==0.10.0
llama-index-retrievers-bm25==0.7.1
llama-index-llms-openai==0.7.5
llama-index-llms-huggingface
qdrant-client==1.17.1
elasticsearch==8.19.3

# 한국어
mecab-python3==1.0.12
python-mecab-ko==1.3.7

# API 클라이언트
openai==2.30.0

# 평가
ragas==0.4.3
langsmith==0.7.24
beir==2.2.0
datasets>=3.4.1,<4.4.0
evaluate==0.4.3

# 문서 파싱
pypdf==4.3.1
python-docx==1.1.2
beautifulsoup4==4.12.3
lxml==5.3.0

# 유틸
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.5
python-dotenv==1.0.1
pyyaml==6.0.2
loguru==0.7.2
rich==13.8.1
jsonlines==4.0.0
scikit-learn==1.7.2
diskcache==5.6.3

# SFT 학습
peft==0.18.1
bitsandbytes==0.49.2
trl<1.0.0
unsloth==2026.4.4
unsloth-zoo==2026.4.4
wandb==0.25.1

# Cohere Reranker API 대안 (선택)
llama-index-postprocessor-cohere-rerank==0.8.0
